Location Rule Cleanup
Hidden in the Depths is triggered by having Ancient Debris in your inventory, not Netherite Scrap.
Cover Me in Debris and Serious Dedication currently have incorrect logic rules that don't align with their physical access. This accounts for that so that they're only shown out of logic when their physical access conditions are met, and only shown in logic when their logical access conditions are met.